refactor: extract cloud sync provider config to src/utils/cloudSyncProviders.js
