+++ src/App.jsx
@@ -8,6 +8,7 @@ import { voiceParseSystemPrompt, voiceParseUserPrompt, taskSuggestSystemPrompt, 
 import { gatherTrmnlData, pushToTrmnl, TRMNL_MARKUP_FULL, TRMNL_MARKUP_HALF_HORIZONTAL, TRMNL_MARKUP_HALF_VERTICAL, TRMNL_MARKUP_QUADRANT } from './trmnl.js';
 import { checkForUpdate } from './versionCheck.js';
 import { getStorageUsage, formatBytes } from './utils/storage.js';
+import { cloudSyncProviders } from './utils/cloudSyncProviders.js';
 
 // Encode a string that may contain non-ASCII characters as Base64.
 // btoa() throws InvalidCharacterError for codepoints > 255 (CJK, emoji, etc.).
@@ -833,141 +834,6 @@ const DailyNotesModal = ({ dateStr, note, onSave, onClose, darkMode, isMobile, t
 // Cloud sync provider abstraction
 // Routes a WebDAV HTTP request through the native Android HTTP bridge when
 // running inside the app (no CORS, no proxy server needed), or through the
-// /api/webdav-proxy/ server-side proxy when running as a PWA/web app.
-const webdavFetch = async (method, targetUrl, authHeaders, body, extraHeaders = {}) => {
-  if (typeof window !== 'undefined' && window.DayGlanceNative?.httpRequest) {
-    // On Android: call the target URL directly with a standard Authorization header.
-    const headers = { ...extraHeaders };
-    if (authHeaders['X-WebDAV-Auth']) {
-      headers['Authorization'] = authHeaders['X-WebDAV-Auth'];
-    } else {
-      Object.assign(headers, authHeaders);
-    }
-    if (body !== undefined && body !== null) {
-      headers['Content-Type'] = extraHeaders['Content-Type'] || 'application/octet-stream';
-    }
-    const result = nativeHttpRequest(method, targetUrl, headers, body ?? '');
-    if (!result) throw new Error('Native HTTP bridge unavailable');
-    return {
-      status: result.status,
-      ok: result.ok,
-      statusText: result.error || '',
-      json: async () => JSON.parse(result.body),
-      text: async () => result.body,
-    };
-  }
-  // On web: route through the server-side CORS proxy.
-  return fetch(`/api/webdav-proxy/?url=${targetUrl}`, {
-    method,
-    headers: { ...authHeaders, ...extraHeaders },
-    ...(body !== undefined && body !== null ? { body } : {}),
-  });
-};
-
-const cloudSyncProviders = {
-  nextcloud: {
-    name: 'Nextcloud / WebDAV',
-    getFileUrl: (config) =>
-      `${config.nextcloudUrl.replace(/\/+$/, '')}/remote.php/dav/files/${encodeURIComponent(config.username)}/dayglance/dayglance-sync.json`,
-    getDirUrl: (config) =>
-      `${config.nextcloudUrl.replace(/\/+$/, '')}/remote.php/dav/files/${encodeURIComponent(config.username)}/dayglance/`,
-    getAuthHeaders: (config) => ({
-      'X-WebDAV-Auth': 'Basic ' + toBase64(config.username + ':' + config.appPassword)
-    }),
-    async upload(config, data) {
-      const fileUrl = this.getFileUrl(config);
-      const dirUrl = this.getDirUrl(config);
-      const authHeaders = this.getAuthHeaders(config);
-      const body = JSON.stringify(data);
-
-      const doUpload = () =>
-        webdavFetch('PUT', fileUrl, authHeaders, body, { 'Content-Type': 'application/json' });
-
-      let res = await doUpload();
-      if (res.status === 404 || res.status === 409) {
-        // Directory doesn't exist, create it
-        await webdavFetch('MKCOL', dirUrl, authHeaders);
-        res = await doUpload();
-      }
-      if (!res.ok) throw new Error(`Upload failed: ${res.status} ${res.statusText}`);
-      return true;
-    },
-    async download(config) {
-      const fileUrl = this.getFileUrl(config);
-      const authHeaders = this.getAuthHeaders(config);
-
-      const res = await webdavFetch('GET', fileUrl, authHeaders);
-
-      if (res.status === 404) return null; // No remote file yet
-      if (!res.ok) throw new Error(`Download failed: ${res.status} ${res.statusText}`);
-      return res.json();
-    },
-    async test(config) {
-      const dirUrl = this.getDirUrl(config);
-      const authHeaders = this.getAuthHeaders(config);
-
-      const res = await webdavFetch('PROPFIND', dirUrl, authHeaders, undefined, { 'Depth': '0' });
-
-      if (res.status === 200 || res.status === 207 || res.status === 404) return { success: true };
-      if (res.status === 401) return { success: false, error: 'Invalid credentials. Check your username and app password.' };
-      return { success: false, error: `Unexpected response: ${res.status}${res.statusText ? ' ' + res.statusText : ''}` };
-    },
-    configFields: [
-      { key: 'nextcloudUrl', label: 'Nextcloud URL', type: 'url', placeholder: 'https://cloud.example.com' },
-      { key: 'username', label: 'Username', type: 'text', placeholder: 'your-username' },
-      { key: 'appPassword', label: 'App Password', type: 'password', placeholder: 'xxxxx-xxxxx-xxxxx-xxxxx-xxxxx' }
-    ],
-    helpText: 'Go to Nextcloud Settings → Security → Devices & sessions → Create new app password'
-  },
-  webdav: {
-    name: 'Generic WebDAV',
-    getFileUrl: (config) =>
-      `${config.webdavUrl.replace(/\/+$/, '')}/dayglance-sync.json`,
-    getDirUrl: (config) =>
-      `${config.webdavUrl.replace(/\/+$/, '')}/`,
-    getAuthHeaders: (config) => ({
-      'X-WebDAV-Auth': 'Basic ' + toBase64(config.username + ':' + config.appPassword)
-    }),
-    async upload(config, data) {
-      const fileUrl = this.getFileUrl(config);
-      const dirUrl = this.getDirUrl(config);
-      const authHeaders = this.getAuthHeaders(config);
-      const body = JSON.stringify(data);
-      const doUpload = () =>
-        webdavFetch('PUT', fileUrl, authHeaders, body, { 'Content-Type': 'application/json' });
-      let res = await doUpload();
-      if (res.status === 404 || res.status === 409) {
-        await webdavFetch('MKCOL', dirUrl, authHeaders);
-        res = await doUpload();
-      }
-      if (!res.ok) throw new Error(`Upload failed: ${res.status} ${res.statusText}`);
-      return true;
-    },
-    async download(config) {
-      const fileUrl = this.getFileUrl(config);
-      const authHeaders = this.getAuthHeaders(config);
-      const res = await webdavFetch('GET', fileUrl, authHeaders);
-      if (res.status === 404) return null;
-      if (!res.ok) throw new Error(`Download failed: ${res.status} ${res.statusText}`);
-      return res.json();
-    },
-    async test(config) {
-      const dirUrl = this.getDirUrl(config);
-      const authHeaders = this.getAuthHeaders(config);
-      const res = await webdavFetch('PROPFIND', dirUrl, authHeaders, undefined, { 'Depth': '0' });
-      if (res.status === 200 || res.status === 207 || res.status === 404) return { success: true };
-      if (res.status === 401) return { success: false, error: 'Invalid credentials. Check your username and password.' };
-      return { success: false, error: `Unexpected response: ${res.status}${res.statusText ? ' ' + res.statusText : ''}` };
-    },
-    configFields: [
-      { key: 'webdavUrl', label: 'WebDAV URL', type: 'url', placeholder: 'https://app.koofr.net/dav/Koofr/dayGLANCE/' },
-      { key: 'username', label: 'Username', type: 'text', placeholder: 'your-username' },
-      { key: 'appPassword', label: 'Password / App Password', type: 'password', placeholder: 'your-password' }
-    ],
-    helpText: 'Enter the full URL of the WebDAV folder where dayGLANCE should store its sync file. Koofr, pCloud, Seafile, and most WebDAV providers are supported.'
-  }
-};
-
 // Cloud sync settings form (extracted to avoid hooks-in-conditional issues)
 const CloudSyncSettingsForm = ({ darkMode, textPrimary, textSecondary, borderClass, hoverBg, cloudSyncConfig, setCloudSyncConfig, cloudSyncTest, provider, currentProvider, onClose, cloudSyncLastSynced }) => {
   const [formData, setFormData] = useState(() => {
